Improved code

+++ Calculators/Number Calculators/All Operations Calculator.cpp
@@ -1,23 +1,31 @@
 // All Operations Calculator by Kazvin
+
 #include <iostream>
 #include <math.h>
 using namespace std;
 
 int main () {
+
     int OperationNumber = 1;
     double FirstNumber, SecondNumber;
+
     while (1) {
+
         cout << "Operation #" << OperationNumber << endl;
-        cout << "First Number: "; cin >> FirstNumber;
-        cout << "Second Number: "; cin >> SecondNumber;
+        cout << "First Number: ";
+        cin >> FirstNumber;
+        cout << "Second Number: ";
+        cin >> SecondNumber;
+
         double Sum = FirstNumber + SecondNumber;
         double Difference = FirstNumber - SecondNumber;
         double Product = FirstNumber * SecondNumber;
         double Quotient = FirstNumber / SecondNumber;
-        int IntQuotient = FirstNumber / SecondNumber;
+        long long IntQuotient = FirstNumber / SecondNumber;
         double Remainder = FirstNumber - (IntQuotient * SecondNumber);
         double Power = pow(FirstNumber, SecondNumber);
         double NthRoot = pow(FirstNumber, 1/SecondNumber);
+
         cout << endl << "Sum: " << Sum << endl;
         cout << "Difference: " << Difference << endl;
         cout << "Product: " << Product << endl;
@@ -25,7 +33,10 @@ int main () {
         cout << "Remainder: " << Remainder << endl;
         cout << "Power: " << Power << endl;
         cout << "Nth Root: " << NthRoot << endl;
+
         cout << "____________________" << endl << endl;
         OperationNumber++;
+
     }
+
 }
